feat (frontend): Organised project structure
- Added new folders for components and pages
- Extracted `HeroSection`, `BackgroundImage` and `ContactForm` components from `Contact` page
- Updated `Contact` page to use the new components
- Moved `Contact` page to the new pages folder
