+++ src/components/BackgroundImage.tsx
@@ -1,0 +1,17 @@
+const BackgroundImage = () => {
+  return (
+    <div
+      className="fixed inset-24 rounded-full"
+      style={{
+        width: "80vw",
+        height: "160vh",
+        backgroundImage: "url('/images/logo_ksi.svg')",
+        backgroundColor: "rgba(242, 242, 242, 0.3)",
+        backgroundSize: "cover",
+        backgroundRepeat: "no-repeat",
+        backgroundPosition: "center",
+      }}
+    />
+  );
+};
+export default BackgroundImage;
